Move TestException into general util package
This is to allow TestException to be used in other tests.

Also refactor how the metrics tests include util packages to simplify
the Deployment methods.
